messaging_app/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0
drf-nested-routers>=0.93.0
Pillow>=9.0.0
python-decouple>=3.8 